Conflicts:
	Source/CNTK/ModelEditLanguage.cpp
	Source/CNTK/SynchronousExecutionEngine.cpp
	Source/ComputationNetworkLib/InputAndParamNodes.h

+++ Source/Readers/ReaderLib/SampleModePacker.cpp
@@ -15,12 +15,11 @@ SampleModePacker::SampleModePacker(
     MemoryProviderPtr memoryProvider,
     TransformerPtr transformer,
     size_t minibatchSize,
-    const std::vector<StreamDescriptionPtr>& streams)
-    : m_transformer(transformer),
-    m_minibatchSize(minibatchSize),
-    m_outputStreams(streams),
-    m_minibatchLayout(std::make_shared<MBLayout>()),
-    m_memoryProvider(memoryProvider)
+    const std::vector<StreamDescriptionPtr>& streams) : m_transformer(transformer),
+                                                        m_minibatchSize(minibatchSize),
+                                                        m_outputStreams(streams),
+                                                        m_minibatchLayout(std::make_shared<MBLayout>()),
+                                                        m_memoryProvider(memoryProvider)
 {
     m_inputStreams = m_transformer->GetStreamDescriptions();
     assert(m_inputStreams.size() == m_outputStreams.size());
@@ -136,7 +135,7 @@ void SampleModePacker::CopySequenceToBuffer(size_t sampleIndex, size_t streamInd
     }
     else
     {
-        RuntimeError("Storage type %d is not supported.", m_inputStreams[streamIndex]->m_storageType);
+        RuntimeError("Storage type %d is not supported.", (int)m_inputStreams[streamIndex]->m_storageType);
     }
 }
